clarify loginName, displayName, userId migrate userId to OwncloudClient
