+++ src/main/java/com/owncloud/android/lib/resources/trashbin/model/TrashbinFile.java
@@ -24,7 +24,6 @@
 
 package com.owncloud.android.lib.resources.trashbin.model;
 
-import android.net.Uri;
 import android.os.Parcel;
 import android.os.Parcelable;
 
@@ -91,7 +90,7 @@ public class TrashbinFile implements Parcelable, Serializable, ServerFileInterfa
         }
         
         fullRemotePath = path;
-        remotePath = fullRemotePath.replace("/trashbin/" + Uri.encode(userId) + "/trash", "");
+        remotePath = fullRemotePath.replace("/trashbin/" + userId + "/trash", "");
         
         setMimeType(we.getContentType());
